Make borgerPC scripts fail on Kiosk and Kiosk scripts fail on borgerPC

See merge request os2borgerpc/os2borgerpc-scripts!202

+++ os2borgerpc/desktop/xdg_policy_numlock.sh
@@ -34,6 +34,11 @@ POLICY="/etc/xdg/autostart/os2borgerpc-numlock.desktop"
 
 set -ex
 
+if get_os2borgerpc_config os2_product | grep --quiet kiosk; then
+  echo "Dette script er ikke designet til at blive anvendt på en kiosk-maskine."
+  exit 1
+fi
+
 if [ "$1" = "" ] || [ "$1" = "false" ] || [ "$1" = "falsk" ]; then
     rm -f "$POLICY"
 else
