+++ os2borgerpc_kiosk/os2borgerpc_kiosk/chromium_delete_profile.sh
@@ -1,3 +1,8 @@
 #!/bin/bash
 
+if ! get_os2borgerpc_config os2_product | grep --quiet kiosk; then
+  echo "Dette script er ikke designet til at blive anvendt på en regulær OS2borgerPC-maskine."
+  exit 1
+fi
+
 rm -r /home/chrome/snap/chromium/common/chromium/Default
